requirements.txt
fastapi
vellox
google-genai
httpx
google-cloud-vision 
pypdf 
python-multipart 
uvicorn[standard] 
google-cloud-storage